Add overloads for intermediate form; move overloads
